app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "google-genai>=1.15.0",
    "jinja2>=3.1.6",
    "logfire>=3.16.0",
    "pydantic>=2.11.4",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.2",
]
